+++ src/TrendyolSharp.Marketplace/Models/Base/Claim.cs
@@ -12,7 +12,7 @@ public sealed class Claim
   public string CargoTrackingLink { get; set; }
   public string CargoSenderNumber { get; set; }
   public string CargoProviderName { get; set; }
-  public int OrderShipmentPackageId { get; set; }
+  public long OrderShipmentPackageId { get; set; }
   public ReplacementInfo ReplacementOutboundpackageinfo { get; set; }
   public RejectedInfo RejectedPackageInfo { get; set; }
   public List<ClaimItem> Items { get; set; }
