+++ src/TrendyolSharp.Marketplace/Models/Base/ClaimItemDetail.cs
@@ -3,7 +3,7 @@
 public class ClaimItemDetail
 {
   public string Id { get; set; }
-  public int OrderLineItemId { get; set; }
+  public long OrderLineItemId { get; set; }
   public ClaimItemReason CustomerClaimItemReason { get; set; }
   public ClaimItemReason TrendyolClaimItemReason { get; set; }
   public ClaimItemStatus ClaimItemStatus { get; set; }
