+++ src/TrendyolSharp.Marketplace/Models/Base/Line.cs
@@ -2,14 +2,14 @@
 
 public sealed class Line
 {
-  public int Quantity { get; set; }
-  public int SalesCampaignId { get; set; }
+  public long Quantity { get; set; }
+  public long SalesCampaignId { get; set; }
   public string ProductSize { get; set; }
   public string MerchantSku { get; set; }
   public string ProductName { get; set; }
-  public int ProductCode { get; set; }
+  public long ProductCode { get; set; }
   public string ProductOrigin { get; set; }
-  public int MerchantId { get; set; }
+  public long MerchantId { get; set; }
   public decimal Amount { get; set; }
   public decimal Discount { get; set; }
   public decimal TyDiscount { get; set; }
@@ -17,7 +17,7 @@ public sealed class Line
   public List<FastDeliveryOption> FastDeliveryOptions { get; set; }
   public string CurrencyCode { get; set; }
   public string ProductColor { get; set; }
-  public int Id { get; set; }
+  public long Id { get; set; }
   public string Sku { get; set; }
   public decimal VatBaseAmount { get; set; }
   public string Barcode { get; set; }
