+++ src/TrendyolSharp.Marketplace/Models/Filter/FilterGetCategoryTree.cs
@@ -2,8 +2,8 @@
 
 public sealed class FilterGetCategoryTree
 {
-  public int? Id { get; set; }
-  public int? ParentId { get; set; }
+  public long? Id { get; set; }
+  public long? ParentId { get; set; }
   public string Name { get; set; }
   public bool? SubCategories { get; set; }
 }
