+++ src/TrendyolSharp.Marketplace/Models/Request/RequestClaimIssue.cs
@@ -3,7 +3,7 @@
 public sealed class RequestClaimIssue
 {
   public string ClaimId { get; set; }
-  public int ClaimIssueReasonId { get; set; }
+  public long ClaimIssueReasonId { get; set; }
   public string ClaimItemIdList { get; set; }
   public string Description { get; set; }
 }
