+++ src/TrendyolSharp.Marketplace/Models/Request/RequestCreateClaim.cs
@@ -5,9 +5,9 @@ namespace TrendyolSharp.Marketplace.Models.Request;
 public sealed class RequestCreateClaim
 {
   public List<CreateClaimItem> ClaimItems { get; set; }
-  public int CustomerId { get; set; }
+  public long CustomerId { get; set; }
   public bool ExcludeListing { get; set; }
   public bool ForcePackageCreation { get; set; }
   public string OrderNumber { get; set; }
-  public int ShipmentCompanyId { get; set; }
+  public long ShipmentCompanyId { get; set; }
 }
