+++ src/TrendyolSharp.Marketplace/Models/Request/RequestCreateCommonLabel.cs
@@ -3,6 +3,6 @@
 public sealed class RequestCreateCommonLabel
 {
   public string Format { get; set; } // required
-  public int BoxQuantity { get; set; }
+  public long BoxQuantity { get; set; }
   public double VolumetricHeight { get; set; }
 }
