+++ src/TrendyolSharp.Marketplace/Models/Response/ResponseGetClaims.cs
@@ -5,8 +5,8 @@ namespace TrendyolSharp.Marketplace.Models.Response;
 
 public sealed class ResponseGetClaims
 {
-  public int TotalElements { get; set; }
-  public int TotalPages { get; set; }
+  public long TotalElements { get; set; }
+  public long TotalPages { get; set; }
   public int Page { get; set; }
   public int Size { get; set; }
   public List<Claim> Content { get; set; }
